refactor(F2-04): consolidate HandleOrderTerminationService into handler
HandleOrderTerminationService had no callers outside of HandleOrderTerminationHandler,
making it an unnecessary indirection. All domain logic (findAndUnlockPosition,
releaseMargin, buildMarginRelease, mapToReleaseReason) is now inlined as private
methods in the handler. Removed the service class and its bean from RunnerConfig.
