+++ spring-application/src/main/java/com/marmitt/application/spring/config/core/RunnerConfig.java
@@ -5,7 +5,6 @@ import com.marmitt.core.application.usecase.portfolio.HandleExecutionConfirmedSe
 import com.marmitt.core.application.usecase.portfolio.HandleMarginReleaseService;
 import com.marmitt.core.application.usecase.portfolio.ReserveCapitalService;
 import com.marmitt.core.application.usecase.portfolio.ReleaseMarginService;
-import com.marmitt.core.application.usecase.runner.HandleOrderTerminationService;
 import com.marmitt.core.application.usecase.runner.PortfolioContextBuilder;
 import com.marmitt.core.application.usecase.runner.ProcessTradeSignalService;
 import com.marmitt.core.application.usecase.runner.RunnerLifecycleUseCase;
@@ -79,16 +78,6 @@ public class RunnerConfig {
             OrderDispatchPort orderDispatchPort
     ) {
         return new ProcessTradeSignalService(runnerRepository, reserveCapitalService, orderDispatchPort);
-    }
-
-    // ── Serviços do fluxo HandleOrderTermination (F2-03) ─────────────────────
-
-    @Bean
-    public HandleOrderTerminationService handleOrderTermination(
-            StrategyRunnerRepositoryPort runnerRepository,
-            ReleaseMarginService releaseMarginService
-    ) {
-        return new HandleOrderTerminationService(runnerRepository, releaseMarginService);
     }
 
     // ── Serviços do fluxo Runner Lifecycle (F2-04) ────────────────────────────
